Fix for when motd doesn't exist

+++ .zshrc.d/99-motd.zsh
@@ -1,5 +1,7 @@
-cmp -s $HOME/.hushlogin /etc/motd
-if [ $? != 0 ]; then
-    tee $HOME/.hushlogin < /etc/motd
-    echo -n "Press Enter to continue: " && read ans
+if [ -f /etc/motd ]; then
+    cmp -s $HOME/.hushlogin /etc/motd
+    if [ $? != 0 ]; then
+        tee $HOME/.hushlogin < /etc/motd
+        echo -n "Press Enter to continue: " && read ans
+    fi
 fi
